Change usize to u32 where appropriate
Potentially could be reduced even further to u16 in a lot of cases
but u32 is safe while still saving a lot of space. In cases where
the numbers are used for indexing and such I left in usize (even
though generally u32 would also be fine) for convenience purposes.